pipelines/pipecat-agent/requirements.txt
pipecatcloud
pipecat-ai[daily,google,krisp,silero]
fastapi
uvicorn
python-dotenv
twilio
